docs: address content type parser review
