+++ lib/content-type.js
@@ -67,6 +67,8 @@ const cache = new Lru(100)
 
 /**
  * ContentType parses and represents the value of the content-type header.
+ * Keep its public structural contract in `FastifyParsedContentType` at
+ * `types/content-type-parser.d.ts` synchronized with these getters.
  *
  * @see https://httpwg.org/specs/rfc9110.html#media.type
  * @see https://httpwg.org/specs/rfc9110.html#parameter
