+++ test/content-parser.test.js
@@ -767,3 +767,95 @@ test('content-type fail when not a valid type', async t => {
     t.assert.equal(error.message, 'The content type should be a string or a RegExp')
   }
 })
+
+const malformedJson = 'application/json,application/json'
+
+function recoverMalformedJson (defaultParser) {
+  return function parseContentType (headerValue) {
+    return defaultParser(headerValue === malformedJson ? 'application/json' : headerValue)
+  }
+}
+
+test('custom content-type parser canonicalizes requests consistently', async t => {
+  const fastify = Fastify({
+    contentTypeHeaderParserFactory: recoverMalformedJson
+  })
+  t.after(() => fastify.close())
+
+  fastify.register((instance, _options, done) => {
+    instance.post('/', {
+      schema: {
+        body: {
+          content: {
+            'application/json': {
+              schema: {
+                type: 'object',
+                required: ['foo'],
+                properties: {
+                  foo: { type: 'string' }
+                }
+              }
+            }
+          }
+        }
+      }
+    }, request => ({
+      body: request.body,
+      mediaType: request.mediaType
+    }))
+    done()
+  })
+
+  const accepted = await fastify.inject({
+    method: 'POST',
+    url: '/',
+    headers: { 'content-type': malformedJson },
+    payload: JSON.stringify({ foo: 'bar' })
+  })
+  t.assert.strictEqual(accepted.statusCode, 200)
+  t.assert.deepStrictEqual(accepted.json(), {
+    body: { foo: 'bar' },
+    mediaType: 'application/json'
+  })
+
+  const rejectedBySchema = await fastify.inject({
+    method: 'POST',
+    url: '/',
+    headers: { 'content-type': malformedJson },
+    payload: JSON.stringify({ bar: 'baz' })
+  })
+  t.assert.strictEqual(rejectedBySchema.statusCode, 400)
+  t.assert.strictEqual(rejectedBySchema.json().code, 'FST_ERR_VALIDATION')
+})
+
+test('custom content-type parser delegates ordinary invalid values', async t => {
+  const fastify = Fastify({
+    contentTypeHeaderParserFactory: recoverMalformedJson
+  })
+  t.after(() => fastify.close())
+  fastify.post('/', () => 'unreachable')
+
+  const response = await fastify.inject({
+    method: 'POST',
+    url: '/',
+    headers: { 'content-type': 'invalid-content-type' },
+    payload: 'body'
+  })
+
+  t.assert.strictEqual(response.statusCode, 415)
+  t.assert.strictEqual(response.json().code, 'FST_ERR_CTP_INVALID_MEDIA_TYPE')
+})
+
+test('contentTypeHeaderParserFactory must be a function', t => {
+  t.assert.throws(
+    () => Fastify({ contentTypeHeaderParserFactory: 'invalid' }),
+    new TypeError("contentTypeHeaderParserFactory option should be a function, instead got 'string'")
+  )
+})
+
+test('contentTypeHeaderParserFactory must return a function', t => {
+  t.assert.throws(
+    () => Fastify({ contentTypeHeaderParserFactory: () => 'invalid' }),
+    new TypeError('contentTypeHeaderParserFactory must return a function')
+  )
+})
